camerabin2: Fix some racyness in tests
Fix some racyness as the test was checking the idle property
right after it got the preview message for video recordings. In some
conditions, it might happen that camerabin2 still hasn't decremented
the processing counter after posting the preview and/or the image/video-done
message and the test will get idle=false and fail.

The approach for checking for the idle property was a busy loop with
a sleep. Far from elegant, but good enough for these tests.