configure: only check for gst-plugins-{good,ugly,ffmpeg} in uninstalled setups
We don't install .pc files for plugin modules, because we only need
them in uninstalled setups, so we can find the plugins for unit tests
(even when run a couple of directories deeper or out-of-tree during
make distcheck).

Try to avoid confusion, so only check for those if this is in fact
an uninstalled setup. For installed setups there's no problem finding
the plugins, we assume they are all in the plugindir specified by
GStreamer core's .pc file.

https://bugzilla.gnome.org/show_bug.cgi?id=668311